+++ shivas-server/src/main/java/org/shivas/server/database/ShivasDatabaseModule.java
@@ -35,8 +35,8 @@ public class ShivasDatabaseModule extends AbstractModule {
         bind(new TypeLiteral<BaseEntityRepository<Integer, Guild>>(){}).to(GuildRepository.class);
         bind(new TypeLiteral<EntityRepository<Integer, Guild>>(){}).to(GuildRepository.class);
 
-        bind(new TypeLiteral<BaseEntityRepository<Long, GuildMember>>(){}).to(GuildMemberRepository.class);
-        bind(new TypeLiteral<EntityRepository<Long, GuildMember>>(){}).to(GuildMemberRepository.class);
+        bind(new TypeLiteral<BaseEntityRepository<Integer, GuildMember>>(){}).to(GuildMemberRepository.class);
+        bind(new TypeLiteral<EntityRepository<Integer, GuildMember>>(){}).to(GuildMemberRepository.class);
 	}
 
 }
